+++ docs/changelog.rst
@@ -15,6 +15,10 @@ Changelog
 * Fixed compilation on systems where OpenSSL's ``ec.h`` header is not
   available, such as CentOS.
 * Added :class:`~cryptography.hazmat.primitives.kdf.pbkdf2.PBKDF2HMAC`.
+* Added :class:`~cryptography.hazmat.primitives.kdf.hkdf.HKDF`.
+* Added :doc:`/hazmat/backends/multibackend`.
+* Set default random for the :doc:`/hazmat/backends/openssl` to the OS random engine.
+* Added :class:`~cryptography.hazmat.primitives.ciphers.algorithms.CAST5` (CAST-128) support.
 
 0.1 - 2014-01-08
 ~~~~~~~~~~~~~~~~